tests/wpt/web-platform-tests/css/css-sizing/intrinsic-size/parsing/intrinsic-size-sets-computed-dimensions.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Intrinsic Size Test: intrinsic-size sets component dimensions.</title>
<link rel="author" title="Vladimir Levin" href="mailto:vmpstr@chromium.org">
<link rel="help" href="https://drafts.csswg.org/css-sizing-4/#intrinsic-size-override">
<meta name="assert" content="intrinsic-size sets component dimensions.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/computed-testcommon.js"></script>
<style>
#horizontal-target {
  writing-mode: horizontal-tb;
}
#vertical-target {
  writing-mode: vertical-rl;
}
</style>
</head>
<body>
<div id=horizontal-target></div>
<div id=vertical-target></div>
<script>

const horizontal = document.getElementById("horizontal-target");
const vertical = document.getElementById("vertical-target");

function runTest() {
  const horizontal_style = getComputedStyle(horizontal);
  const vertical_style = getComputedStyle(vertical);
  test(() => {
    assert_equals(horizontal_style["intrinsic-width"], "1px");
    assert_equals(vertical_style["intrinsic-width"], "1px");
  }, 'intrinsic-width set correctly');

  test(() => {
    assert_equals(horizontal_style["intrinsic-height"], "2px");
    assert_equals(vertical_style["intrinsic-height"], "2px");
  }, 'intrinsic-height set correctly');

  test(() => {
    assert_equals(horizontal_style["intrinsic-inline-size"], "1px");
    assert_equals(vertical_style["intrinsic-inline-size"], "2px");
  }, 'intrinsic-inline-size set correctly');

  test(() => {
    assert_equals(horizontal_style["intrinsic-block-size"], "2px");
    assert_equals(vertical_style["intrinsic-block-size"], "1px");
  }, 'intrinsic-block-size set correctly');
}

window.onload = () => {
  horizontal.style["intrinsic-size"] = "1px 2px";
  vertical.style["intrinsic-size"] = "1px 2px";
  runTest();
};
</script>
</body>
</html>
